requirements.txt
PyJWT==2.10.1
python-dotenv==1.1.1
